Moved graph client to it's own file
